Add design-specification/ and .png files within
I have once again added a directory containing files which detail the
design specification to be achieved in the landing page. This time, I
was able to download the pages directly as .png files, which have not
compromised the quality.

As before, full-design.png details the overall design specification to
be achieved in the project, and color-and-font-styles.png describes
specific information about the colors and fonts to be used in the
landing page.